change to Ubuntu/trusty65 and new config

+++ shell/setup.sh
@@ -1,4 +1,5 @@
 #! /bin/bash
+
 # Set timezone
 echo "Europe/Kiev" | sudo tee /etc/timezone
 sudo dpkg-reconfigure --frontend noninteractive tzdata
@@ -14,16 +15,39 @@ sudo chmod 644 /home/vagrant/.gitconfig
 sudo chmod 644 /home/vagrant/.profile
 sudo chmod 644 /home/vagrant/.selected_editor
 
+#install mysql server
+if [ ! -f /var/log/mysql.setup ];
+then
 echo mysql-server mysql-server/root_password password root | sudo debconf-set-selections
 echo mysql-server mysql-server/root_password_again password root | sudo debconf-set-selections
+sudo apt-get install -y mysql-server mysql-client
+fi
 
-sudo apt-get install -y git-core curl wget mc
+#install soft
+if [ ! -f /var/log/soft.setup ];
+then
+sudo apt-get install -y git-core curl wget mc atop htop
+sudo touch /var/log/soft.install
+fi
+
+#install apache2
+if [ ! -f /var/log/soft.setup ];
+then
 sudo apt-get install -y apache2
 sudo a2enmod rewrite
-sudo apt-get install -y mysql-server mysql-client
+sudo touch /var/log/apache2.install
+fi
+
+#install php
+if [ ! -f /var/log/php.install ];
+then
 sudo apt-get install -y php5 libapache2-mod-php5 php5-cli php5-mysql php5-curl php5-gd php5-mcrypt php-pear  php5-xdebug
+sudo touch /var/log/php.install
+fi
 
 #install phpmyadmin
+if [ ! -f /var/log/phpmyadmin.install ];
+then
     echo 'phpmyadmin phpmyadmin/dbconfig-install boolean false' | debconf-set-selections
 	echo 'phpmyadmin phpmyadmin/reconfigure-webserver multiselect apache2' | debconf-set-selections
 
@@ -40,45 +64,73 @@ sudo apt-get install -y php5 libapache2-mod-php5 php5-cli php5-mysql php5-curl p
 	echo 'dbconfig-common dbconfig-common/app-password-confirm password root' | debconf-set-selections
 	echo 'dbconfig-common dbconfig-common/password-confirm password root' | debconf-set-selections
 sudo apt-get install -y phpmyadmin
+sudo touch /var/log/phpmyadmin.install
+fi
 
 #install Mailcather
-sudo apt-get install -y ruby1.9.1-dev
-sudo apt-get install -y sqlite3 libsqlite3-dev
-sudo apt-get install -y ruby rubygems
-gem install mailcatcher
-
-# Install xdebug
-if [ ! -f /var/log/xdebugsetup ];
+if [ ! -f /var/log/mailcather.install ];
 then
-    sudo pecl install xdebug
-    XDEBUG_LOCATION=$(find / -name 'xdebug.so' 2> /dev/null)
-
-    sudo touch /var/log/xdebugsetup
+sudo apt-get install -y build-essential libsqlite3-dev ruby1.9.3
+sudo gem install mailcatcher
+sudo touch /var/log/mailcather.install
 fi
 
 #Install Composer
+if [ ! -f /var/log/composer.install ];
+then
 sudo curl -sS https://getcomposer.org/installer | php
 sudo mv composer.phar /usr/local/bin/composer
+sudo touch /var/log/composer.install
+fi
+
+#CONFIGURATION
 
 # Configure PHP
-if [ ! -f /var/log/phpsetup ];
+if [ ! -f /var/log/php.setup ];
 then
-    sudo sed -i '/;sendmail_path =/c sendmail_path = "/usr/local/bin/catchmail"' /etc/php5/apache2/php.ini
-    sudo sed -i '/;sendmail_path =/c sendmail_path = "/usr/local/bin/catchmail"' /etc/php5/cli/php.ini
+    # Configure sendmail path to mailcathcer
+    sudo sed -i '/;sendmail_path =/c sendmail_path = "/usr/local/bin/catchmail -f local@dev"' /etc/php5/apache2/php.ini
+    sudo sed -i '/;sendmail_path =/c sendmail_path = "/usr/local/bin/catchmail -f local@dev"' /etc/php5/cli/php.ini
+    # Display  Errors
     sudo sed -i '/display_errors = Off/c display_errors = On' /etc/php5/apache2/php.ini
+    sudo sed -i '/display_errors = Off/c display_errors = On' /etc/php5/cli/php.ini
     sudo sed -i '/error_reporting = E_ALL & ~E_DEPRECATED/c error_reporting = E_ALL | E_STRICT' /etc/php5/apache2/php.ini
+    sudo sed -i '/error_reporting = E_ALL & ~E_DEPRECATED/c error_reporting = E_ALL | E_STRICT' /etc/php5/cli/php.ini
     sudo sed -i '/html_errors = Off/c html_errors = On' /etc/php5/apache2/php.ini
-    sudo sed -i '/upload_max_filesize = 2M/c upload_max_filesize = 16M' /etc/php5/apache2/php.ini
-     echo "zend_extension='$XDEBUG_LOCATION'" | sudo tee -a /etc/php5/apache2/php.ini > /dev/null
-     echo "zend_extension='$XDEBUG_LOCATION'" | sudo tee -a /etc/php5/cli/php.ini > /dev/null
-    sudo touch /var/log/phpsetup
+    sudo sed -i '/html_errors = Off/c html_errors = On' /etc/php5/apache2/php.ini
+
+    sudo sed -i '/log_errors = Off/c log_errors = On' /etc/php5/apache2/php.ini
+    sudo sed -i '/log_errors = Off/c log_errors = On' /etc/php5/cli/php.ini
+
+    sudo sed -i '/upload_max_filesize = 2M/c upload_max_filesize = 64M' /etc/php5/apache2/php.ini
+    sudo sed -i '/upload_max_filesize = 2M/c upload_max_filesize = 64M' /etc/php5/cli/php.ini
+
+    sudo sed -i '/post_max_size = 8M/c post_max_size = 64M' /etc/php5/apache2/php.ini
+    sudo sed -i '/post_max_size = 8M/c post_max_size = 64M' /etc/php5/cli/php.ini
+
+    sudo sed -i '/;error_log = php_errors.log/c error_log = /var/log/php_errors.log' /etc/php5/apache2/php.ini
+    sudo sed -i '/;error_log = php_errors.log/c error_log = /var/log/php_errors.log' /etc/php5/apache2/php.ini
+
+    sudo touch /var/log/php.setup
+fi
+
+# Configure Apache2
+if [ ! -f /var/log/apache2.setup ];
+then
+    sudo sed -i 's/AllowOverride None/AllowOverride all/' /etc/apache2/apache2.conf
+    sudo sed -i 's/export APACHE_RUN_USER=www-data/export APACHE_RUN_USER=vagrant/' /etc/apache2/envvars
+    sudo sed -i 's/export APACHE_RUN_GROUP=www-data/export APACHE_RUN_GROUP=vagrant/' /etc/apache2/envvars
+    # If you want to install a different path to the web directory. For example:
+    #sudo sed -i 's/DocumentRoot \/var\/www/DocumentRoot \/mnt\/var\/www\/html/g' /etc/apache2/sites-available/default
+    #sudo sed -i 's/<Directory \/var\/www\/>/<Directory \/mnt\/var\/www\/html\/>/' /etc/apache2/sites-available/default
+    sudo touch /var/log/apache2.setup
 fi
 
 #restart apache2
 sudo service apache2 restart
 
 #ran Mailcather
-mailcatcher --http-ip=11.11.11.11
+mailcatcher --http-ip=10.10.10.10
